adds approve request to store
