+++ src/components/ApproveRequests.vue
@@ -1,13 +1,13 @@
 /* eslint-disable vue/no-use-v-if-with-v-for */
 <template>
-  <Loading v-if="loading"/>
+  <Loading v-if="this.$store.state.loading"/>
     <ul
       class="campaigns"
       v-for="campaign in userContributions"
       v-bind:key="campaign.id"
       >
       <request-list
-        v-if="!loading"
+        v-if="!this.$store.state.loading"
         v-bind:campaign="campaign"
         v-on:approval-count="approvalCount"
         v-on:approve-request="approveRequest"
@@ -18,10 +18,8 @@
 <script>
 
 // import Loading from './Loading.vue';
-import axios from "axios";
 import RequestList from "@/ui/RequestContainer.vue"
 import Campaign from "../contracts/campaign";
-import { VUE_APP_API_URL } from "../env";
 import Loading from "./Loading.vue";
 
 export default {
@@ -43,7 +41,8 @@ export default {
   },
   data() {
     return {
-      loading: false
+      loading: false,
+      error: ''
     }
   },
   components: {
@@ -55,31 +54,12 @@ export default {
       this.$store.dispatch('getUserContribution', this.accountNum)
     },
     /* eslint-disable */
-    async approveRequest({ address, ethId, requestID }) {
-      this.loading = true;
-      const campaignInstance = await Campaign.at(address)
-      console.log(campaignInstance)
-      let result;
+    approveRequest({ address, ethId, requestID }) {
+      const payload = { address, ethId, requestID }
       try {
-        result = await campaignInstance.approveRequest(ethId, { from: this.$store.state.accountNum})
-      } catch(error) {
-        console.log(error)
-        this.loading = false;
-      }
-      const request = await campaignInstance.requests(ethId)
-      const approvalCount = parseInt(request.approvalCount.toString())
-      const voterCount = await campaignInstance.approversCount()
-      const voterString = voterCount.toString();
-      console.log('approvalCount', approvalCount)
-      console.log('voterCount', voterCount)
-      if (result && (approvalCount >= parseInt(voterString) / 2)) {
-        const resp = await axios.put(`${VUE_APP_API_URL}requests/${requestID}`, {
-          approved: "true"
-        });
-        console.log(resp);
-      }
-      if(result) {
-        this.loading = false;
+        this.$store.dispatch('approveRequest', payload)
+      } catch (error) {
+        this.error = { error }
       }
     },
     async approvalCount({address, ethId}) {
